backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.1
numpy>=1.21.0
scikit-image>=0.18.0
pydantic>=1.8.2
python-jose>=3.3.0
sqlalchemy>=1.4.23
passlib>=1.7.4
opencv-python>=4.5.3
pytest>=6.2.5
httpx>=0.18.2
pywavelets>=1.2.0 